+++ CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.3
+- Changed to execute the `view` command by default, when no command argument is supplied
+- Migrated from `tui` to `ratatui` and bumped dependencies
+- Added `worker millis per event` chart on the pipeline's plugin details
