+++ Cargo.toml
@@ -7,30 +7,29 @@ documentation = "https://github.com/edmocosta/tuistash"
 keywords = ["logstash", "tui", "cli", "terminal"]
 categories = ["command-line-utilities"]
 authors = ["Edmo Costa <edinhow@gmail.com>"]
-version = "0.1.1"
+version = "0.1.3"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.3.0", features = ["std", "derive", "help", "usage", "suggestions"], default-features = false }
+clap = { version = "4.3.21", features = ["std", "derive", "help", "usage", "suggestions"], default-features = false }
 ureq = { version = "2.7.1", features = ["json", "tls"], default-features = false }
 serde = { version = "1.0", features = ["derive"], default-features = false }
-serde_with = { version = "3.0.0", default-features = false }
+serde_with = { version = "3.2.0", default-features = false }
 serde_json = { version = "1.0", default-features = false }
-tabled = { version = "0.10", features = ["color"], default-features = false }
-owo-colors = { version = "3.5.0", default-features = false }
-json_to_table = { version = "*", default-features = false }
-base64 = { version = "0.21.0", features = ["std"], default-features = false }
-rustls = { version = "0.21.2", features = ["dangerous_configuration"], default-features = false }
-colored_json = { version = "3", default-features = false }
-chrono = { version = "0.4.26", features = ["std", "default"], default-features = false }
-humansize = { version = "2.1.3", default-features = false }
-humantime = { version = "2.1.0" }
-tui = { version = "0.19.0", features = ["crossterm"] }
-crossterm = { version = "0.26.1", default-features = false, features = ["event-stream"] }
-num-format = { version = "0.4.4", default-features = false, features = ["with-num-bigint"] }
+tabled = { version = "0.14.0", features = ["std","default"] }
+colored = { version = "2.0.4", default-features = false }
+json_to_table = { version = "0.5.0", default-features = false }
+base64 = { version = "0.21", features = ["std"], default-features = false }
+rustls = { version = "0.21", features = ["dangerous_configuration"], default-features = false }
+colored_json = { version = "3.2", default-features = false }
+humansize = { version = "2.1", default-features = false }
+humantime = { version = "2.1" }
+ratatui = { version = "0.22.0", features = ["crossterm"] }
+crossterm = { version = "0.27.0", default-features = false, features = ["event-stream"] }
+num-format = { version = "0.4", default-features = false, features = ["with-num-bigint"] }
 human_format = { version = "1.0" }
-uuid = { version = "1.3.3", features = ["v4"] }
-
+uuid = { version = "1.4", features = ["v4"] }
+time = { version = "0.3", features = ["default", "formatting", "local-offset"] }
 
 [[bin]]
 name = "tuistash"
