+++ src/cli/commands/formatter.rs
