+++ src/cli/commands/view/ui_node_tab.rs
@@ -1,19 +1,86 @@
-use humansize::{format_size_i, ToF64, DECIMAL};
-use tui::backend::Backend;
-use tui::layout::{Constraint, Direction, Layout, Rect};
-use tui::style::{Color, Modifier, Style};
-use tui::text::Span;
-use tui::widgets::{Axis, Block, Borders, Chart, Dataset, Gauge, GraphType};
-use tui::{symbols, Frame};
+use humansize::{format_size_i, DECIMAL};
+use ratatui::backend::Backend;
+use ratatui::layout::{Constraint, Direction, Layout, Rect};
+use ratatui::style::{Color, Modifier, Style};
+use ratatui::text::{Line, Span};
+use ratatui::widgets::{Axis, Block, Borders, Chart, Dataset, Gauge, GraphType, Paragraph, Wrap};
+use ratatui::{symbols, Frame};
 
+use crate::commands::formatter::{DurationFormatter, NumberFormatter};
 use crate::commands::view::app::App;
 use crate::commands::view::charts::{
-    create_binary_size_label_spans, create_percentage_label_spans, create_timestamp_label_spans,
-    DEFAULT_LABELS_COUNT,
+    create_chart_binary_size_label_spans, create_chart_percentage_label_spans,
+    create_chart_timestamp_label_spans, DEFAULT_LABELS_COUNT,
 };
-use crate::commands::view::flow_charts::{render_flow_chart, render_plugins_flow_chart};
+use crate::commands::view::flow_metrics_charts::{
+    draw_flow_metric_chart, draw_plugin_throughput_flow_chart,
+};
 
-pub(crate) fn render_node_charts<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+pub(crate) fn draw_node_tab<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+where
+    B: Backend,
+{
+    let chunks = Layout::default()
+        .constraints([Constraint::Percentage(100)].as_ref())
+        .split(area);
+
+    draw_node_widgets(f, app, chunks[0]);
+}
+
+fn draw_node_widgets<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+where
+    B: Backend,
+{
+    let chunks = Layout::default()
+        .constraints(vec![Constraint::Length(3), Constraint::Percentage(82)])
+        .direction(Direction::Vertical)
+        .split(area);
+    {
+        // Node overview
+        let events_block = Block::default().title("Overview").borders(Borders::ALL);
+        let overview_text: Line = if let Some(node_stats) = &app.state.node_stats {
+            Line::from(vec![
+                Span::styled("Events in: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.events.r#in.format_number()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled("Events out: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.events.out.format_number()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled("Reloads: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.reloads.successes.format_number()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled("Pipeline workers: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.pipeline.workers.to_string()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled(
+                    "Pipeline batch size: ",
+                    Style::default().fg(Color::DarkGray),
+                ),
+                Span::from(node_stats.pipeline.batch_size.to_string()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled("Version: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.version.as_str()),
+                Span::styled(" | ", Style::default().fg(Color::Yellow)),
+                Span::styled("Uptime: ", Style::default().fg(Color::DarkGray)),
+                Span::from(node_stats.jvm.uptime_in_millis.format_duration()),
+            ])
+        } else {
+            Line::from(vec![Span::styled(
+                "-",
+                Style::default().fg(Color::DarkGray),
+            )])
+        };
+
+        let info_paragraph = Paragraph::new(overview_text)
+            .block(events_block)
+            .wrap(Wrap { trim: true });
+        f.render_widget(info_paragraph.clone(), chunks[0]);
+
+        draw_node_charts(f, app, chunks[1]);
+    }
+}
+
+fn draw_node_charts<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
 where
     B: Backend,
 {
@@ -33,14 +100,14 @@ where
         )
         .split(main_chunks[0]);
 
-    render_process_file_descriptor_gauge(f, app, flow_chart_chunks[0]);
-    render_plugins_flow_chart(
+    draw_process_file_descriptor_gauge(f, app, flow_chart_chunks[0]);
+    draw_plugin_throughput_flow_chart(
         f,
         "Throughput",
         &app.state.chart_flow_plugins_throughput,
         flow_chart_chunks[1],
     );
-    render_flow_chart(
+    draw_flow_metric_chart(
         f,
         "Queue Backpressure",
         None,
@@ -59,12 +126,12 @@ where
         )
         .split(main_chunks[1]);
 
-    render_jvm_heap_chart(f, app, node_chart_chunks[0]);
-    render_jvm_non_heap_chart(f, app, node_chart_chunks[1]);
-    render_process_cpu_chart(f, app, node_chart_chunks[2]);
+    draw_jvm_heap_chart(f, app, node_chart_chunks[0]);
+    draw_jvm_non_heap_chart(f, app, node_chart_chunks[1]);
+    draw_process_cpu_chart(f, app, node_chart_chunks[2]);
 }
 
-fn render_jvm_heap_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+fn draw_jvm_heap_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
 where
     B: Backend,
 {
@@ -72,8 +139,8 @@ where
     let mut heap_used_data: Vec<(f64, f64)> = vec![];
 
     for data in &app.state.chart_jvm_heap_state.data_points {
-        heap_max_data.push((data.timestamp.to_f64(), data.heap_max_in_bytes.to_f64()));
-        heap_used_data.push((data.timestamp.to_f64(), data.heap_used_in_bytes.to_f64()));
+        heap_max_data.push((data.timestamp as f64, data.heap_max_in_bytes as f64));
+        heap_used_data.push((data.timestamp as f64, data.heap_used_in_bytes as f64));
     }
 
     let current_max_heap = app
@@ -121,7 +188,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_jvm_heap_state.x_axis_bounds())
-                .labels(create_timestamp_label_spans(
+                .labels(create_chart_timestamp_label_spans(
                     app.state
                         .chart_jvm_heap_state
                         .x_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -131,7 +198,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_jvm_heap_state.y_axis_bounds())
-                .labels(create_binary_size_label_spans(
+                .labels(create_chart_binary_size_label_spans(
                     app.state
                         .chart_jvm_heap_state
                         .y_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -141,7 +208,7 @@ where
     f.render_widget(chart, area);
 }
 
-fn render_jvm_non_heap_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+fn draw_jvm_non_heap_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
 where
     B: Backend,
 {
@@ -149,13 +216,10 @@ where
     let mut non_heap_committed: Vec<(f64, f64)> = vec![];
 
     for data in &app.state.chart_jvm_non_heap_state.data_points {
-        non_heap_used.push((
-            data.timestamp.to_f64(),
-            data.non_heap_used_in_bytes.to_f64(),
-        ));
+        non_heap_used.push((data.timestamp as f64, data.non_heap_used_in_bytes as f64));
         non_heap_committed.push((
-            data.timestamp.to_f64(),
-            data.non_heap_committed_in_bytes.to_f64(),
+            data.timestamp as f64,
+            data.non_heap_committed_in_bytes as f64,
         ));
     }
 
@@ -207,7 +271,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_jvm_non_heap_state.x_axis_bounds())
-                .labels(create_timestamp_label_spans(
+                .labels(create_chart_timestamp_label_spans(
                     app.state
                         .chart_jvm_non_heap_state
                         .x_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -217,7 +281,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_jvm_non_heap_state.y_axis_bounds())
-                .labels(create_binary_size_label_spans(
+                .labels(create_chart_binary_size_label_spans(
                     app.state
                         .chart_jvm_non_heap_state
                         .y_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -227,7 +291,7 @@ where
     f.render_widget(chart, area);
 }
 
-fn render_process_cpu_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+fn draw_process_cpu_chart<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
 where
     B: Backend,
 {
@@ -236,7 +300,7 @@ where
         .chart_process_cpu
         .data_points
         .iter()
-        .map(|p| (p.timestamp.to_f64(), p.percent.to_f64()))
+        .map(|p| (p.timestamp as f64, p.percent as f64))
         .collect();
 
     let current_value = app
@@ -265,7 +329,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_process_cpu.x_axis_bounds())
-                .labels(create_timestamp_label_spans(
+                .labels(create_chart_timestamp_label_spans(
                     app.state
                         .chart_process_cpu
                         .x_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -275,7 +339,7 @@ where
             Axis::default()
                 .style(Style::default().fg(Color::Gray))
                 .bounds(*app.state.chart_process_cpu.y_axis_bounds())
-                .labels(create_percentage_label_spans(
+                .labels(create_chart_percentage_label_spans(
                     app.state
                         .chart_process_cpu
                         .y_axis_labels_values(DEFAULT_LABELS_COUNT),
@@ -285,7 +349,7 @@ where
     f.render_widget(chart, area);
 }
 
-fn render_process_file_descriptor_gauge<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
+fn draw_process_file_descriptor_gauge<B>(f: &mut Frame<B>, app: &mut App, area: Rect)
 where
     B: Backend,
 {
